Add new Shell classes and modify existing ones
Added new classes to handle the creation and management of shells. This includes a ShellBlueprint class, an ElsaBuilder class and an ApplicationBlueprint class. Modifications were also made to existing shell-related classes to accommodate the newly added classes. Changes were also implemented in the 'Elsa' project to include reference to the 'Elsa.Tenants' project.